requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt
#
astroid==2.15.6
    # via pylint
bandit==1.7.5
    # via cs.aws-account (setup.py)
boto3==1.28.15
    # via cs.aws-account (setup.py)
botocore==1.31.15
    # via
    #   boto3
    #   cs.aws-account (setup.py)
    #   s3transfer
cachetools==5.3.1
    # via cs.aws-account (setup.py)
coverage==7.2.7
    # via cs.aws-account (setup.py)
cs-ratelimit==1.3.0
    # via cs.aws-account (setup.py)
dill==0.3.7
    # via pylint
flake8==6.1.0
    # via cs.aws-account (setup.py)
gitdb==4.0.10
    # via gitpython
gitpython==3.1.32
    # via bandit
isort==5.12.0
    # via pylint
jinja2==3.1.2
    # via cs.aws-account (setup.py)
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy-object-proxy==1.9.0
    # via astroid
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mdurl==0.1.2
    # via markdown-it-py
pbr==5.11.1
    # via stevedore
platformdirs==3.10.0
    # via pylint
pycodestyle==2.11.0
    # via flake8
pydocstyle==6.3.0
    # via cs.aws-account (setup.py)
pyflakes==3.1.0
    # via flake8
pygments==2.15.1
    # via rich
pylint==2.17.5
    # via cs.aws-account (setup.py)
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0.1
    # via
    #   bandit
    #   cs.aws-account (setup.py)
rich==13.5.0
    # via bandit
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
stevedore==5.1.0
    # via bandit
tomlkit==0.12.1
    # via pylint
urllib3==1.26.16
    # via botocore
wrapt==1.15.0
    # via astroid
zope-component==6.0
    # via
    #   cs-ratelimit
    #   cs.aws-account (setup.py)
    #   zope-security
zope-configuration==5.0
    # via cs.aws-account (setup.py)
zope-event==5.0
    # via
    #   zope-component
    #   zope-schema
zope-exceptions==5.0.1
    # via zope-testrunner
zope-hookable==5.4
    # via zope-component
zope-i18nmessageid==6.0.1
    # via
    #   zope-configuration
    #   zope-security
zope-interface==6.0
    # via
    #   cs-ratelimit
    #   cs.aws-account (setup.py)
    #   zope-component
    #   zope-configuration
    #   zope-exceptions
    #   zope-location
    #   zope-proxy
    #   zope-schema
    #   zope-security
    #   zope-testrunner
zope-location==5.0
    # via zope-security
zope-proxy==5.0.0
    # via
    #   zope-location
    #   zope-security
zope-schema==7.0.1
    # via
    #   cs-ratelimit
    #   cs.aws-account (setup.py)
    #   zope-configuration
    #   zope-location
    #   zope-security
zope-security==6.1
    # via cs.aws-account (setup.py)
zope-testrunner==6.0
    # via cs.aws-account (setup.py)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
